switch to C++17
* switch to CGAL 6.0.1 which can use boost.multiprecision and gmp/mpfr
* CGAL 6.0.1 now works also with cfs on Windows (w/o gmp/mpfr)
* also upgrade gmp and mpfr which now runs on macOS with M1
* macOS/Linux still requires gmp/mpfr to be linked (bug in CGAL)
* cfsdat still does not compile on Windows (issue with Point.h)
* switch to flann 1.9.2 because of C++17
* add lz4 compression required by flann
* solve small warning issues

+++ source/unittests/unittests.cc
@@ -425,7 +425,7 @@ BOOST_AUTO_TEST_CASE(signed_omp_loop)
   Vector<double> v(100, 2.0);
 
   #pragma omp parallel for
-  for(int i = 0; i < v.GetSize(); i++)
+  for(int i = 0; i < (int) v.GetSize(); i++)
     v[i] = std::exp(v[i]);
 
   #pragma omp parallel for
